+++ src/StephJob/Core/Domain/Naics2Digit.cs
@@ -1,19 +1,13 @@
 ﻿namespace StephJob.Core.Domain
 {
-    public class Naics
+    public class Naics2Digit
     {
+        public int Naics2DigitId { get; set; }
+
         public int NaicsId { get; set; }
 
-        public int Level { get; set; }
+        public string NaicsCode2Digits { get; set; }
 
-        public string Hierarchical { get; set; }
-
-        public int Code { get; set; }
-
-        public string Title { get; set; }
-
-        public string Superscript { get; set; }
-
-        public string Definition { get; set; }
+        public string Sector { get; set; }
     }
 }
