+++ src/StephJob/Core/Repositories/IGreenGoodServiceRepository.cs
@@ -1,0 +1,8 @@
+﻿namespace StephJob.Core.Repositories
+{
+    using StephJob.Core.Domain;
+
+    public interface IGreenGoodServiceRepository : IRepository<GreenGoodService>
+    {
+    }
+}
