+++ src/StephJob/wwwroot/datasets/Green goods and services (US).csv
@@ -1,0 +1,15 @@
+Industry,NaicsCode2Digit,GgsEmployment,GgsPercent,TotalEmployment
+Natural resources and mining,"11","63344",3.5,"1798593"
+Utilities,22,"69031",12.5,"551287"
+Construction,23,"385777",7,"5489499"
+Manufacturing,31-33,"492985",4.3,"11487496"
+Trade,44-45,"205567",1,"19947787"
+Transportation and warehousing,48-49,"242137",6.1,"3943659"
+Information,51,"33321",1.2,"2703886"
+Financial activities,52,462,0,"7401812"
+"Professional, scientific, and technical services",54,"355386",4.8,"7457913"
+Management of companies and enterprises,55,"62630",3.4,"1854778"
+Administrative and waste services,56,"330650",4.5,"7399320"
+Education and health services,61,"28789",0.2,"18656159"
+Leisure and hospitality,71,"20642",0.2,"13006814"
+"Other services, except public administration",81,"51841",1.2,"4349563"
